Refactor shellcheck source reposet.inc.sh
In order to be able to run shellcheck correctly when the sources are not
stored in a directory named 'src', e.g. after installation into the
directory /usr/local/bin, change the shellcheck source pragmas in the
scripts to not contain src/. Instead, cd into the src/directory when
running shellcheck.